fix: 사진용 커뮤니티로 들어오는 경로 수정
