+++ components/MapMarker/MapList.jsx
@@ -39,7 +39,7 @@ export const MapList = ({ route, navigation }) => {
       cacheTime: 1000 * 60 * 60 * 24,
     });
 
-    navigation.navigate('PhotoCom', {
+    navigation.navigate('MapPhotoCom', {
       photo: photo,
     });
   };
