+++ package.json
@@ -43,6 +43,7 @@
     "react-native-modal": "^13.0.1",
     "react-native-pager-view": "6.2.0",
     "react-native-reanimated": "~3.3.0",
+    "react-native-reanimated-zoom": "^0.3.3",
     "react-native-root-toast": "^3.5.1",
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0",
