+++ screens/MyPage/Map/Map.jsx
@@ -69,7 +69,7 @@ export const Map = ({ navigation }) => {
       cacheTime: 1000 * 60 * 60 * 24,
     });
 
-    navigation.navigate('PhotoCom', {
+    navigation.navigate('MapPhotoCom', {
       photo: photo,
     });
   };
