+++ screens/MyPage/index.js
@@ -1,6 +1,7 @@
 export { Post } from './Post/Post';
 export { Circle } from './Circle/Circle';
 export { Map } from './Map/Map';
-export { PhotoCom } from './Circle/PhotoCom';
+export { MapPhotoCom } from './Circle/MapPhotoCom';
+export { CirclePhotoCom } from './Circle/CirclePhotoCom';
 export { SingleCircle } from './Circle/SingleCircle';
 export { SplashUI } from './Circle/SplashUI';
